securitycenter/v1beta/python/google-cloud-modelarmor/v0/requirements.txt
google-cloud-modelarmor==0.2.8
